requirements.txt
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
idna==2.9
invoke==1.4.1
pika==1.1.0
pycodestyle==2.5.0
python-logstash==0.4.6
requests==2.23.0
robotframework==3.1.2
robotframework-seleniumlibrary==4.3.0
selenium==3.141.0
urllib3==1.25.8
wrapt==1.12.0
structlog==20.1.0
python-json-logger==0.1.11